refactor: [app] move login related plugins
